+++ src/bin/tui/widgets/tracing_window.rs
@@ -82,7 +82,7 @@ impl<'a> TracingWindow<'a> {
                 let unproven = self.theme.unproven_requirement;
                 let skipped = self.theme.skipped_requirement;
                 for i in requirements {
-                    let goal = &i.task.goal().term;
+                    let goal = i.task.goal().term();
                     result.push(Line::from(match i.status {
                         SolutionStatus::Answer(_) => Span::styled(format!("  ☑  {goal}"), proven),
                         SolutionStatus::Err(_) => Span::styled(format!("  ☒  {goal}"), unproven),